Refactor: (#17) 서비스 응답 포맷 간소화 및 JWT/쿠키 설정 환경변수화
- AuthService는 처리만, AuthController에서 응답 통일
- JWT·쿠키 설정은 .env 관리
- 인증 가드 분리해 예외 처리 명확화
