+++ src/auth/auth.controller.ts
@@ -12,7 +12,12 @@ export class AuthController {
 
   @Post('signup')
   async signup(@Body() signupRequestDto: SignupRequestDto) {
-    return await this.authService.signup(signupRequestDto);
+    await this.authService.signup(signupRequestDto);
+    return {
+      status: 'success',
+      message: '회원가입에 성공했습니다.',
+      data: null,
+    };
   }
 
   @Post('login')
@@ -20,21 +25,22 @@ export class AuthController {
     @Body() loginRequestDto: LoginRequestDto,
     @Res({ passthrough: true }) res: Response,
   ) {
-    const result = await this.authService.login(loginRequestDto);
+    const { accessToken, refreshToken } =
+      await this.authService.login(loginRequestDto);
 
-    res.cookie('refresh_token', result.data.refreshToken, {
+    res.cookie('refresh_token', refreshToken, {
       httpOnly: true,
       secure: true,
       sameSite: 'strict',
       maxAge: 7 * 24 * 60 * 60 * 1000,
     });
 
-    const { refreshToken, ...restData } = result.data;
-
     return {
-      status: result.status,
-      message: result.message,
-      data: restData,
+      status: 'success',
+      message: '로그인에 성공했습니다.',
+      data: {
+        accessToken,
+      },
     };
   }
 }
