+++ src/auth/guards/custom-jwt-auth.guard.ts
@@ -1,0 +1,27 @@
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+import { TokenExpiredError, JsonWebTokenError } from 'jsonwebtoken';
+
+@Injectable()
+export class CustomJwtAuthGuard extends AuthGuard('jwt') {
+  handleRequest(err: any, user: any, info: any) {
+    if (err || !user) {
+      if (
+        err instanceof TokenExpiredError ||
+        info?.name === 'TokenExpiredError'
+      ) {
+        throw new UnauthorizedException(
+          '토큰이 만료되었습니다. 다시 로그인해주세요.',
+        );
+      } else if (
+        err instanceof JsonWebTokenError ||
+        info?.name === 'JsonWebTokenError'
+      ) {
+        throw new UnauthorizedException('유효하지 않은 토큰입니다.');
+      } else {
+        throw new UnauthorizedException('인증에 실패했습니다.');
+      }
+    }
+    return user;
+  }
+}
